// apps/admin/src/lib/api/generated/models/partnerPipelineOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.50
 */
import type { PartnerPipelineOutOperation } from './partnerPipelineOutOperation';

export interface PartnerPipelineOut {
  attempts: number;
  available_at: string;
  created_at: string;
  error: string | null;
  external_id: string | null;
  finished_at: string | null;
  id: string;
  operation: PartnerPipelineOutOperation;
  parent_id: string | null;
  product_id: string | null;
  product_name: string | null;
  provider: string;
  status: string;
}